[refactoring] package 분리
