+++ src/main/java/step2/RacingGameResultModel.java
@@ -2,6 +2,8 @@ package step2;
 
 import java.util.List;
 
+import step2.game.domain.Cars;
+
 public class RacingGameResultModel {
     private final List<Cars> cars;
