+++ src/main/java/step2/game/domain/CarFactory.java
@@ -1,4 +1,4 @@
-package step2;
+package step2.game.domain;
 
 public interface CarFactory {
