+++ src/main/java/step2/game/domain/DefaultMoveStrategy.java
@@ -1,4 +1,4 @@
-package step2;
+package step2.game.domain;
 
 public class DefaultMoveStrategy implements MoveStrategy {
 }
