+++ src/main/java/step2/game/domain/MoveStrategy.java
@@ -1,4 +1,6 @@
-package step2;
+package step2.game.domain;
+
+import step2.game.util.RacingGameRandomUtils;
 
 public interface MoveStrategy {
     default boolean isMove() {
