+++ src/main/java/step2/game/domain/RacingGameRecord.java
@@ -1,4 +1,4 @@
-package step2;
+package step2.game.domain;
 
 import java.util.ArrayList;
 import java.util.List;
