+++ src/test/java/step2/CarsTest.java
@@ -3,6 +3,11 @@ package step2;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
+import step2.game.domain.Cars;
+import step2.game.domain.DefaultCarFactoryImpl;
+import step2.game.domain.DefaultMoveStrategy;
+import step2.game.domain.MoveStrategy;
+
 import static org.assertj.core.api.Assertions.*;
 
 public class CarsTest {
