+++ src/test/java/step2/RacingGameRecordTest.java
@@ -4,6 +4,9 @@ import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
+import step2.game.domain.Cars;
+import step2.game.domain.RacingGameRecord;
+
 import static org.assertj.core.api.Assertions.*;
 
 public class RacingGameRecordTest {
